Add files via upload
Here are the two sets of data I have. I think initially we are all going to work on NM_UFO_DATA_ALL because that contains the data specifically for New Mexico, but as we get more comfortable maybe we can branch out to a broader analysis to include the complete data set referred to as scrubbed.